Fixed more warnings
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@848 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Core/Mesh/ArrayKernelT.cc
@@ -76,7 +76,7 @@ void ArrayKernel::garbage_collection(std_API_Container_VHandlePointer& vh_to_upd
   const bool track_hhandles = ( !hh_to_update.empty() );
   const bool track_fhandles = ( !fh_to_update.empty() );
 
-  int i, i0, i1, nV(n_vertices()), nE(n_edges()), nH(2*n_edges()), nF(n_faces());
+  int i, i0, i1, nV(int(n_vertices())), nE(int(n_edges())), nH(int(2*n_edges())), nF(int(n_faces()));
 
   std::vector<VertexHandle>    vh_map;
   std::vector<HalfedgeHandle>  hh_map;
@@ -247,9 +247,9 @@ void ArrayKernel::garbage_collection(std_API_Container_VHandlePointer& vh_to_upd
     }
   }
 
-  const int vertexCount   = vertices_.size();
-  const int halfedgeCount = edges_.size() * 2;
-  const int faceCount     = faces_.size();
+  const int vertexCount   = int(vertices_.size());
+  const int halfedgeCount = int(edges_.size() * 2);
+  const int faceCount     = int(faces_.size());
 
   // Update the vertex handles in the vertex handle vector
   typename std_API_Container_VHandlePointer::iterator v_it(vh_to_update.begin()), v_it_end(vh_to_update.end());
